Move configs into rails_helper
